backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
pydantic==2.10.6
python-docx==1.1.2
docxtpl==0.17.0
# Tùy chọn cho module AI (Chỉ cài nếu dùng AI)
openai==1.65.1
google-generativeai==0.8.4
anthropic==0.49.0
